services/ai/requirements.txt
flask==3.0.0
gunicorn==22.0.0
torch==2.6.0
torchvision==0.21.0
transformers==4.48.0
Pillow==10.3.0
numpy==1.26.3
opencv-python-headless==4.9.0.80
plyfile==1.0.3
# ml-sharp is installed via: pip install git+https://github.com/apple/ml-sharp.git
